requirements.txt
streamlit
openai
gtts
pygame
python-dotenv
sounddevice
scipy
faster-whisper
